fix: remove daoselector from ds and change its style
