--- apps/dashboard/shared/components/design-system/navigation/sidebar-dao-dropdown/index.ts
@@ -1,2 +1,0 @@
-export { SidebarDAODropdown } from "@/shared/components/design-system/navigation/sidebar-dao-dropdown/SidebarDAODropdown";
-export type { SidebarDAODropdownProps } from "@/shared/components/design-system/navigation/sidebar-dao-dropdown/SidebarDAODropdown";
